Simplified yaml config loading
Removed the ability to specify a config_name.
Annotated enforced passing a string
String can be valid yaml or a file path.
Will check if file path and if so cast it into a pathlib.Path file which is what ruaml.yaml wants.
Added unit tests for loading config value