Add a simple failing test case.
This test current fails. Either I'm doing it wrong, or --fail-force
argument doesn't work. The test fails because when the check function
fails, the stdin gets and EOF which raises an exception.

The diff view of the log compare can hopefully be improved, because
it doesn't really show the useful details at the moment.