Cleanup of the sequencer code.
No logic changes, just removing some dead code. Parts of this are
referenced in some of the sequencer tests, but they are broken
at the moment anyway. At this point, removing things that aren't
used in the actual logic when running is a bonus to make reading
the code easier.