Continue to expand the script2log integration tests.
There are tests now for all the failed check function cases. Some
need to be marked as XFAIL for now because of an assumption in the
log comparison. Some logs show a bug where the last line is not
the test summary, which the log compare assumes is the case.