Build some conventions around the pubsub message usage.
Remove the "Finish" messsage. This was internal to the GUI and
not necessary. We just pass an appropriate method reference now
to be used as a call back.

Now, the only use of pubsub subscriptions is for connections to the
sequence.

Rename all topic handler to be prefixed with _topic_ and name them
the same as the topic itself. Previously a few topic were subscribed
multiple times. Change that so the mapping is one-to-one.