Refactor the driver import strategy.
This is not 100% where I'd like, but better than it was. General idea is
to move driver specific knowledge out of config/core libraries and into
the drivers themselves.

In the process, the sequence instantiation has been moved to happen only
once the main script is running. Previously this happened at import time.

This all needs testing still, because I have no hardware to test against.